+++ server/tests/test_assignment.py
@@ -1235,9 +1235,7 @@ class TestDynamic:
                 "info": {
                     "assignment": "dynamic",
                     "shuffle": False,
-                    "dynamic_top": 1,
                     "dynamic_warmup": 2,
-                    "dynamic_backoff": 0,
                 },
                 "data": [
                     [{"src": "a", "tgt": {"model1": "b", "model2": "c"}}],
@@ -1270,7 +1268,6 @@ class TestDynamic:
                 "info": {
                     "assignment": "dynamic",
                     "shuffle": False,
-                    "dynamic_top": 1,
                     "dynamic_warmup": 2,
                 },
                 "data": [
@@ -1392,8 +1389,7 @@ class TestDynamic:
                 "info": {
                     "assignment": "dynamic",
                     "docs_per_user": 2,
-                    "dynamic_top": 1,
-                    "dynamic_first": 2,
+                    "dynamic_coldstart": 2,
                 },
                 "data": [
                     [{"src": "a", "tgt": {"model1": "b", "model2": "c"}}],
@@ -1428,8 +1424,7 @@ class TestDynamic:
                 "info": {
                     "assignment": "dynamic",
                     "docs_per_user": 3,
-                    "dynamic_top": 1,
-                    "dynamic_first": 2,
+                    "dynamic_coldstart": 2,
                 },
                 "data": [
                     [{"src": "a", "tgt": {"model1": "b", "model2": "c"}}],